backend/setup/requirements.txt
altgraph==0.17.2
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.4
Django==4.0.6
django-filter==22.1
django-storages==1.12.3
djangorestframework==3.13.1
future==0.18.2
idna==3.3
jdatetime==4.1.0
kavenegar==1.1.2
paramiko==2.11.0
pefile==2021.9.3
Pillow==9.2.0
psycopg2-binary==2.9.3
psycopg2
pycparser==2.21
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
PyNaCl==1.5.0
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
